Continuing with project page
To be able to do a custom project page, is necessary support the
attachment of many images for an unique custom post type. To do that
i found the attachment plugin that enable that in a easy way.

Now, i will do the logic and layout for the project page.
Before that, i did some refactor to make the page.php easier to read
and mantain, separating the parts of each type of page in functions.