Page for a single solution done
Now, we are able to load a single solution
with quality displayed on images attached to the solution content administration.

Besides that, some refactor was made in code.

The theme still verbose, a lot of things should be improved and refactored. But, i will let it for another time. =P